Updating base Taskhub application
Okay, so I've updated the Calculatron flask application to pander to Taskhub. This is a VERY rough mock-up and will be updated as we go along. The login and signup page seems to work, as the user's information is being uploaded to the database and detects if a username appears. Just need to figure out how to redirect the user to the home page and grayout signup when the user is logged in.

On an organizational standpoint, I've moved code snippets into the "Backup" folder if we ever need code snippets or if anything happens to the main application.