Updated Signup page and database
Updated the signup page to take user's last name, username, and phone number.   The user's password will now be hidden and can be shown by an eye icon. I also updated the database to the current layout as well.

I also added the userPage.html to redirect upon signing up or logging in.