Login and signup redirection.
Now, upon the user logging in or signing up a Taskhub account, the header Login and signup links will disappear and be replaced with "Logout". Upon logging out, the user will be redirected to the login page and the header will change back. I added the Lucky Foot logo to the project for later use.