Added Calendar.html + General tweaks
I have implemented jsdelivr code for the calendar on the calendar.html page. I added a calendar link to the header as well for login users. Using Susana's login HTML as a base, I added three tabs: Calendar, Reminders, and Completed Tasks (This is prone to change).

As for general tweaks, I added JavaScript and jQuery to the theme.html to ensure the webpage functioned on mobile devices.